124 - an initial implementation of keyboard support. it covers all of the basic grounds, including raising the expected events, nested containers and allowing mouse drag and drops as it goes. Committing as a good first checkpoint
